+++ src/components/FilterSelectInput/styled.tsx
@@ -1,9 +1,8 @@
 import styled from "styled-components";
-import { colors } from "../../styles";
 import {
   FilterSize,
   FilterVariant,
-  FILTER_VARIANTS,
+  getFilterVariantConfig,
 } from "../FilterInputAbstract/types";
 import { ContextMenu2TextInputItem } from "../ContextMenu2";
 
@@ -27,10 +26,13 @@ export const Select = styled.button<StyledProps>`
   height: 100%;
   padding: 0 8px 0 54px;
   border: 0;
-  background: ${({ $variant }) => FILTER_VARIANTS[$variant].background};
+  background: ${({ $variant, theme }) => {
+    const variantConfig = getFilterVariantConfig(theme);
+    return variantConfig[$variant].background;
+  }};
   outline-offset: -1px;
   text-align: left;
-  color: ${colors.basic[900]};
+  color: ${({ theme }) => theme.palette.black};
   cursor: pointer;
 
   &:focus {
@@ -74,8 +76,102 @@ export const StyledContextMenu2TextInputItem = styled(
 )<StyledContextMenu2TextInputItemProps>`
   input {
     &::placeholder {
-      color: ${({ $placeholderColor }) =>
-        $placeholderColor || colors.basic[600]};
+      color: ${({ $placeholderColor, theme }) =>
+        $placeholderColor || theme.palette.icon.line};
     }
+  }
+`;
+
+export const StyledTrigger = styled.button<{
+  $isOpen?: boolean;
+  $placeholder?: boolean;
+  $placeholderColor?: string;
+  $size: "small" | "medium" | "large";
+}>`
+  position: relative;
+  display: flex;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  border: 0;
+  margin: 0;
+  padding-left: 6px;
+  background: transparent;
+  text-align: left;
+  overflow: hidden;
+  cursor: pointer;
+
+  &:focus {
+    outline: none;
+  }
+
+  /* リストが表示されているときの矢印アニメーション */
+  .arrow-down {
+    transform: ${({ $isOpen }) =>
+      $isOpen ? "rotate(180deg)" : "rotate(0deg)"};
+    transition: transform 0.2s ease;
+  }
+
+  .select-content {
+    display: flex;
+    align-items: center;
+    gap: 6px;
+    overflow: hidden;
+    font-size: 13px;
+    line-height: 16px;
+    color: ${({ theme }) => theme.palette.black};
+  }
+
+  /* プレースホルダー表示 */
+  .placeholder {
+    font-size: 13px;
+    line-height: 16px;
+    font-weight: 400;
+    color: ${({ $placeholderColor, theme }) =>
+      $placeholderColor || theme.palette.icon.line};
+  }
+
+  .label {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+
+  .arrow-container {
+    position: absolute;
+    right: 8px;
+    display: flex;
+    align-items: center;
+    pointer-events: none;
+  }
+`;
+
+export const SelectPanelInner = styled.div`
+  max-height: 180px;
+  overflow: auto;
+  scrollbar-width: none;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
+
+  li {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+  }
+`;
+
+export const SelectPanelContainer = styled.div`
+  background-color: ${({ theme }) => theme.palette.background.default};
+  border-radius: 6px;
+  border: 1px solid #e5e5e5;
+  overflow: hidden;
+  /* Drop shadow Common */
+  box-shadow: 0px 0px 16px rgba(4, 28, 51, 0.08);
+  z-index: 1000;
+
+  ${SelectPanelInner}:not(:last-child) {
+    border-top: 1px solid ${({ theme }) => theme.palette.gray.main};
   }
 `;